Add files via upload
I have gone through the each variable after removing empty rows. There was an issue with the Bag Weight Variable where the units was not same for the value and this variable was string data type. Therefore, I created a user-define function to deal this and make sure all the values had the same units and were also numerical variable.